dom/devicestorage/test/test_usedSpace.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup();

function usedSpaceSuccess(e) {
  ok(e.target.result > 0, "total bytes should exist and be greater than zero");
  devicestorage_cleanup();
}

function usedSpaceError(e) {
  ok(false, "usedSpaceError was called");
  devicestorage_cleanup();
}

var storage = navigator.getDeviceStorage("pictures");

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  devicestorage_cleanup();
}

function addSuccess(e) {
  request = storage.usedSpace();
  ok(request, "Should have a non-null request");

  request.onsuccess = usedSpaceSuccess;
  request.onerror = usedSpaceError;
}

request = storage.addNamed(createRandomBlob('image/png'), "a/b.png");
request.onsuccess = addSuccess;
request.onerror = addError;

</script>
</pre>
</body>
</html>
